+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "super-chat-image-generator",
+  "name": "awesome-image-generators",
   "version": "0.1.0",
-  "homepage": "https://ykicchan.github.io/super-chat-image-generator",
+  "homepage": "https://ykicchan.github.io/awesome-image-generators",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.14.1",
@@ -14,6 +14,7 @@
     "html-to-image": "^1.9.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-router-dom": "6",
     "react-scripts": "5.0.0",
     "sass": "^1.50.0",
     "typescript": "^4.4.2",
