+++ public/404.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <title>Single Page Apps for GitHub Pages</title>
+    <title>Awesome Image Generators</title>
     <script type="text/javascript">
       // Single Page Apps for GitHub Pages
       // MIT License
