+++ public/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <meta name="description" content="Super Chat Image Generator / スパチャ画像ジェネレーター" />
+    <meta name="description" content="Awesome Image Generators / Super Chat / スパチャ画像 / Discord Rank Card" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <!-- Start Single Page Apps for GitHub Pages -->
     <script type="text/javascript">
@@ -31,7 +31,7 @@
       }(window.location))
     </script>
     <!-- End Single Page Apps for GitHub Pages -->
-    <title>Super Chat Image Generator</title>
+    <title>Awesome Image Generators</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
