+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "Super Chat",
-  "name": "Super Chat Image Generator",
+  "short_name": "Image Generators",
+  "name": "Awesome Image Generators",
   "icons": [
     {
       "src": "favicon.ico",
