+++ src/App/SuperChat/Preview/index.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { SuperChat } from 'App';
 import styles from './styles.module.scss';
 
 interface P extends SuperChat {
