+++ src/App/SuperChat/styles.module.css
@@ -1,0 +1,13 @@
+.h1 {
+  margin-top: 0;
+}
+
+.h2 {
+  margin: 1em 0;
+}
+
+.download {
+  margin: 1.5rem auto 0;
+  max-width: 180px;
+  min-width: 120px;
+}
