+++ src/App/SuperChat/super-chat.d.ts
@@ -1,0 +1,6 @@
+declare interface SuperChat {
+  icon: string;
+  name: string;
+  price: number;
+  message: string;
+}
