+++ src/App/styles.module.css
@@ -7,20 +7,6 @@
   margin: 0 auto;
 }
 
-.h1 {
-  margin-top: 0;
-}
-
-.h2 {
-  margin: 1em 0;
-}
-
-.download {
-  margin: 1.5rem auto 0;
-  max-width: 180px;
-  min-width: 120px;
-}
-
 .footer {
   margin-top: auto;
   padding: 2rem 0 1rem;
